test: shared sendTransaction tests for both facades
Extract sendTransaction and sendManyOutputsTransaction
coverage into the shared test framework. Adds
sendTransaction adapter method that handles pinCode
injection for service facade.

- shared/send-transaction.test.ts: internal transfer
  balance invariant, external transfer balance decrease
- fullnode-specific/send-transaction.test.ts: address
  tracking, custom/fee tokens, multisig, sendManyOutputs
- service-specific/send-transaction.test.ts: full structure
  validation, P2SH target, changeAddress verification
