+++ __tests__/integration/walletservice_facade.test.ts
@@ -327,153 +327,8 @@ describe('basic transaction methods', () => {
     }
   });
 
-  describe('sendTransaction - native token', () => {
-    it('should send a simple transaction with native token', async () => {
-      const sendTransaction = await gWallet.sendTransaction(walletWithTxs.addresses[0], 10n, {
-        pinCode,
-      });
-
-      // Shallow validate all properties of the returned Transaction object
-      expect(sendTransaction).toEqual(
-        expect.objectContaining({
-          // Core transaction identification
-          hash: expect.any(String),
-
-          // Inputs and outputs
-          inputs: expect.any(Array),
-          outputs: expect.any(Array),
-
-          // Transaction metadata
-          version: expect.any(Number),
-          weight: expect.any(Number),
-          nonce: expect.any(Number),
-          signalBits: expect.any(Number),
-          timestamp: expect.any(Number),
-
-          // Transaction relationships
-          parents: expect.arrayContaining([expect.any(String)]),
-          tokens: expect.any(Array), // May be empty array
-
-          // Headers
-          headers: expect.any(Array), // May be empty
-        })
-      );
-
-      // Deep validate the Inputs and Outputs arrays
-      expect(sendTransaction.inputs).toEqual(
-        expect.arrayContaining([
-          expect.objectContaining({
-            hash: expect.any(String),
-            index: expect.any(Number),
-            data: expect.any(Buffer),
-          }),
-        ])
-      );
-
-      expect(sendTransaction.outputs).toEqual(
-        expect.arrayContaining([
-          expect.objectContaining({
-            value: expect.any(BigInt),
-            script: expect.any(Buffer),
-            tokenData: expect.any(Number),
-          }),
-        ])
-      );
-
-      // Additional specific validations
-      expect(sendTransaction.hash).toHaveLength(64); // Transaction hash should be 64 hex characters
-      expect(sendTransaction.inputs.length).toBeGreaterThan(0); // Should have at least one input
-      expect(sendTransaction.outputs.length).toBeGreaterThan(0); // Should have at least one output
-      expect(sendTransaction.tokens).toHaveLength(0); // Not populated if only the native token is sent
-      expect(sendTransaction.parents).toHaveLength(2); // Should have exactly 2 parents
-      expect(sendTransaction.timestamp).toBeGreaterThan(0); // Should have a valid timestamp
-
-      // Verify the transaction was sent to the correct address with correct value
-      const recipientOutput = sendTransaction.outputs.find(output => output.value === 10n);
-      expect(recipientOutput).toStrictEqual(
-        expect.objectContaining({
-          value: 10n,
-          tokenData: 0,
-        })
-      );
-
-      await pollForTx(gWallet, sendTransaction.hash!);
-    });
-
-    it('should send a transaction to a P2SH (multisig) address', async () => {
-      // Use a P2SH address from the multisig wallet constants
-      const p2shAddress = WALLET_CONSTANTS.multisig.addresses[0];
-
-      const sendTransaction = await gWallet.sendTransaction(p2shAddress, 5n, {
-        pinCode,
-      });
-
-      // Validate the transaction was created successfully
-      expect(sendTransaction).toEqual(
-        expect.objectContaining({
-          hash: expect.any(String),
-          inputs: expect.any(Array),
-          outputs: expect.any(Array),
-        })
-      );
-
-      // Wait for transaction to be confirmed and verify it on the full node
-      await pollForTx(gWallet, sendTransaction.hash!);
-
-      // Get the full transaction from the network to verify the P2SH output
-      const fullTx = await gWallet.getFullTxById(sendTransaction.hash!);
-      expect(fullTx.success).toBe(true);
-
-      // Find the output with the P2SH address
-      const p2shOutput = fullTx.tx.outputs.find(output => output.decoded?.address === p2shAddress);
-
-      expect(p2shOutput).toBeDefined();
-      expect(p2shOutput!.value).toBe(5n);
-      expect(p2shOutput!.decoded.type).toBe('MultiSig');
-      expect(p2shOutput!.decoded.address).toBe(p2shAddress);
-    });
-
-    it('should send a transaction with a set changeAddress', async () => {
-      ({ wallet } = buildWalletInstance({ words: walletWithTxs.words }));
-      await wallet.start({ pinCode, password });
-
-      const sendTransaction = await wallet.sendTransaction(walletWithTxs.addresses[1], 4n, {
-        pinCode,
-        changeAddress: walletWithTxs.addresses[0],
-      });
-
-      // Verify that the only outputs were the recipient and the change address
-      expect(sendTransaction.outputs.length).toBe(2);
-
-      // Verify the transaction was sent to the correct address with correct value
-      let recipientIndex;
-      let changeIndex;
-      sendTransaction.outputs.forEach((output, index) => {
-        if (output.value === 4n) {
-          recipientIndex = index;
-        } else if (output.value === 6n) {
-          changeIndex = index;
-        }
-      });
-
-      // Confirm the addresses through UTXO queries
-      await pollForTx(wallet, sendTransaction.hash!);
-      const recipientUtxo = await wallet.getUtxoFromId(sendTransaction.hash!, recipientIndex);
-      expect(recipientUtxo).toStrictEqual(
-        expect.objectContaining({
-          address: walletWithTxs.addresses[1],
-          value: 4n,
-        })
-      );
-      const changeUtxo = await wallet.getUtxoFromId(sendTransaction.hash!, changeIndex);
-      expect(changeUtxo).toStrictEqual(
-        expect.objectContaining({
-          address: walletWithTxs.addresses[0],
-          value: 6n,
-        })
-      );
-    });
-  });
+  // sendTransaction - native token tests moved to
+  // shared/send-transaction.test.ts and service-specific/send-transaction.test.ts
 
   describe('createNewToken, getTokenDetails', () => {
     const tokenName = 'TestToken';
